modify structure of celery notification for pubsub/http
