+++ requirements/base.txt
@@ -11,5 +11,5 @@ idna==2.7                 # via requests
 pytz==2018.5              # via django
 requests==2.19.1
 urllib3==1.23             # via requests
-firebase-admin==3.0.0
 celery==4.3.0
+google-cloud-pubsub==1.0.0
